requirements.txt
easydict==1.9
extensions==0.4
h5py==3.7.0
matplotlib==3.5.2
numpy==1.21.6
open3d==0.9.0.0
opencv_python==4.6.0.66
pandas==1.3.5
PyYAML==6.0
scipy==1.7.3
setuptools==63.1.0
torchvision==0.13.0
tqdm==4.64.0
transforms3d==0.3.1
tensorboardX==2.5.1
tensorboard
Ninja